Added all the docker things and have started to move things around a bit
