--- .vscode/settings.json
@@ -1,4 +1,0 @@
-{
-    "python.pythonPath": "/Users/tmikonen/anaconda3/bin/python",
-    "python.linting.pylintEnabled": true
-}
