+++ docker/inventory-system/Dockerfile
@@ -1,0 +1,8 @@
+FROM python:3.11.0-slim-buster as mtg-inventory
+
+WORKDIR /src
+
+COPY /requirements/dev.txt requirements.txt
+
+RUN pip install requirements.txt
+
