+++ requirements/dev.txt
@@ -1,0 +1,7 @@
+numpy==1.23.4
+matplotlib==3.6.1
+shapely==1.8.5.post1
+scipy==1.9.3
+Pillow==9.3.0
+ImageHash==4.3.1
+opencv-python==4.6.0.66
