+++ src/magic_card_detector.py
